requirements.txt
python-dotenv
requests
flask
flask-socketio
eventlet